Rename settings to semantic-settings and use assert rather than throw
